Switch to dual CJS/MJS bundle, node:test, and express 4/5 parallel
- Rollup で dist/express-tee.{cjs,mjs} を生成 (既存 esm/*.mjs 別 build を撤去)
- mocha → node:test
- TypeScript strip-only で .ts 直実行 (lib/test/builder すべて .ts のまま)
- engines: node >= 22
- @types/express4 / @types/express5 alias で express 4/5 並列テスト
- express-intercept dep を ^1.2.0 に bump
- README に "Works with Express 4 and 5" 追記、SYNOPSIS は元から ESM
- LICENSE / README の Copyright を 2026 まで延長
- CI matrix を node-version: [22, 24] に
- バージョン: 1.1.0 → 1.2.0
