+++ test/express-tee.express4.test.ts
@@ -1,0 +1,6 @@
+// Express 4 系列での結合テスト
+
+import express from "express4";
+import {registerSharedTests} from "./lib/shared.ts";
+
+registerSharedTests("express4", express);
