+++ test/express-tee.express5.test.ts
@@ -1,0 +1,6 @@
+// Express 5 系列での結合テスト
+
+import express from "express5";
+import {registerSharedTests} from "./lib/shared.ts";
+
+registerSharedTests("express5", express);
